+++ src/database/schema/sessions.ts
@@ -37,6 +37,9 @@ export const sessions = sqliteTable(
 		/** Temperature used */
 		temperature: real('temperature').notNull(),
 
+		/** The thoughts/reasoning outputted by the LLM, if available */
+		reasoning: text('reasoning'),
+
 		/** The answer given by the LLM */
 		answer: text('answer').notNull(),
 
@@ -126,6 +129,9 @@ export const sessionEvaluations = sqliteTable(
 
 		/** Temperature used */
 		temperature: real('temperature').notNull(),
+
+		// Note: reasoning is not available when generating structured objects, even for reasoning models,
+		// and cannot be stored here as a result.
 
 		/** Whether the evaluation was passed or not */
 		pass: integer('pass').notNull(),