+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.3] - 2025-03-02
+
 ### Added
 
 - Preliminary support for reasoning models (limited to Deepseek-R1 on Ollama) with hardcoded maximum 120 sec and 1500 token limits (this will be configurable in the future).
@@ -48,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 🎉 Version `0.9.0` is out... with all the base features. It isn't quite yet stable or tested enough to be relied upon.
 
 [Unreleased]: https://github.com/gerukin/ai-tester/compare/v0.9.2...HEAD
+[0.9.3]: https://github.com/gerukin/ai-tester/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/gerukin/ai-tester/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/gerukin/ai-tester/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/gerukin/ai-tester/releases/tag/0.9.0